Fix up integration test files
