+++ cli/tests/integration/eval_tests.rs
@@ -64,7 +64,7 @@ itest!(dyn_import_eval {
 // because itest! splits args on whitespace.
 itest!(v8_flags_eval {
   args: "eval --v8-flags=--expose-gc console.log(typeof(gc))",
-  output: "v8_flags.js.out",
+  output: "run/v8_flags.js.out",
 });
 
 itest!(check_local_by_default {
