+++ cli/tests/testdata/cert/cafile_ts_fetch.ts
@@ -1,3 +1,3 @@
-fetch("https://localhost:5545/cafile_ts_fetch.ts.out")
+fetch("https://localhost:5545/cert/cafile_ts_fetch.ts.out")
   .then((r) => r.text())
   .then((t) => console.log(t.trimEnd()));
