+++ cli/tests/testdata/echo.ts
