+++ cli/tests/testdata/run/event_listener_error.ts
