+++ cli/tests/testdata/run/event_listener_error_handled.ts
