+++ cli/tests/testdata/run/event_listener_error_immediate_exit.ts.out
