+++ cli/tests/testdata/run/event_listener_error_immediate_exit_worker.ts
