+++ cli/tests/testdata/run/no_prompt.ts
